requirements.txt

numpy
jupyter
ipython
tqdm
torchvision
seaborn
matplotlib
opencv-python
scikit-learn
captum
scikit-image
captum-rise

